requirements.txt
streamlit==1.36.0
openai==1.40.3
python-dotenv==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
rich==13.7.1
httpx==0.27.0
pydantic==2.7.4
streamlit-local-storage==0.0.8